Test works after picking changes from stash
