+++ client/main.go
@@ -69,7 +69,6 @@ const (
 var status State
 
 var startTime time.Time
-var demoTaken = false
 var leftTime, servTime time.Duration
 
 // timeTaken per question displayed on the top.
@@ -86,118 +85,14 @@ var sessionId string
 
 var infoPage InformationPage
 
-func setupInfoPage(th, tw int) {
-	instructions = termui.NewPar("")
-	instructions.BorderLabel = "Instructions"
-	instructions.Height = 50
-	instructions.Width = tw
-	instructions.PaddingTop = 2
-
-	infoPage.terminal = termui.NewPar(`
-		- Please ensure that you can see all the 4 borders of the Instructions box.
-		- If you can't see them, you need to increase the size of your terminal or adjust the font-size to a smaller value.
-		- DO NOT proceed with the test, until you are able to see all 4 outer borders of the Instructions box.`)
-	infoPage.terminal.BorderLabel = "Terminal"
-	infoPage.terminal.Height = 8
-	infoPage.terminal.Width = tw
-	infoPage.terminal.PaddingLeft = 2
-
-	// TODO - Take duration from constant.
-	infoPage.general = termui.NewPar(`
-		- By taking this test, you agree not to discuss/post the questions shown here.
-		- The duration of the test is 60 mins. Timing would be clearly shown.
-		- Once you start the test, the timer would not stop, irrespective of any client side issues.
-		- Questions can have single or multiple correct answers. They will be shown accordingly.
-		- Your total score and the time left at any point in the test would be displayed on the top.
-		- You would be given the option to have a second attempt at a question if your first answer is wrong.
-		- The scoring for each attempt of a question, would be visible to you in a separate section.
-		- At point you can press Ctrl-q to end the test.`)
-	infoPage.general.BorderLabel = "General"
-	infoPage.general.Height = 15
-	infoPage.general.Width = tw
-	infoPage.general.PaddingLeft = 2
-
-	infoPage.scoring = termui.NewPar(`
-		- There is NEGATIVE scoring for wrong answers. So, please DO NOT GUESS.
-		- If you skip a question, the score awarded is always ZERO.
-		- You might be given the option to recover from negative score with a second attempt.
-		- In the above case, please note that another wrong answer would have further negative score.
-		- scoring would be clearly marked in the question on the right hand side box.`)
-	infoPage.scoring.BorderLabel = "Scoring"
-	infoPage.scoring.Height = 10
-	infoPage.scoring.Width = tw
-	infoPage.scoring.PaddingLeft = 2
-
-	infoPage.contact = termui.NewPar(`
-		- If there are any problems with the setup, or something is unclear, please DO NOT start the test.
-		- Send email to contact@dgraph.io and tell us the problem. So we can solve it before you take the test.`)
-	infoPage.contact.BorderLabel = "Contact"
-	infoPage.contact.Height = 10
-	infoPage.contact.Width = tw
-	infoPage.contact.PaddingLeft = 2
-
-	infoPage.demo = termui.NewPar("We have a demo of the how the test would look like. Press s to start the demo.")
-	infoPage.demo.Border = false
-	infoPage.demo.Height = 3
-	infoPage.demo.Width = tw
-	infoPage.demo.TextFgColor = termui.ColorCyan
-	infoPage.demo.PaddingLeft = 2
-	infoPage.demo.PaddingTop = 1
-}
-
-var qp QuestionsPage
-
-func setupQuestionsPage() {
-	qp.timeLeft = termui.NewPar(fmt.Sprintf("%02d:00", testDur))
-	qp.timeLeft.Height = 3
-	qp.timeLeft.BorderLabel = "Time Left"
-
-	qp.timeSpent = termui.NewPar("00:00")
-	qp.timeSpent.Height = 3
-	qp.timeSpent.BorderLabel = "Time spent"
-
-	ts := 00.0
-	qp.score = termui.NewPar(fmt.Sprintf("%2.1f", ts))
-	qp.score.BorderLabel = "Total Score"
-	qp.score.Height = 3
-
-	qp.lastScore = termui.NewPar("0.0")
-	qp.lastScore.BorderLabel = "Last Score"
-	qp.lastScore.Height = 3
-
-	qp.que = termui.NewPar("")
-	qp.que.BorderLabel = "Question"
-	qp.que.PaddingLeft = 1
-	qp.que.PaddingRight = 1
-	qp.que.PaddingBottom = 1
-	qp.que.Height = 33
-
-	qp.scoringInfo = termui.NewPar("")
-	qp.scoringInfo.BorderLabel = "Scoring"
-	qp.scoringInfo.PaddingTop = 1
-	qp.scoringInfo.PaddingLeft = 1
-	qp.scoringInfo.Height = 33
-
-	qp.answers = termui.NewPar("")
-	qp.answers.TextFgColor = termui.ColorCyan
-	qp.answers.BorderLabel = "Answers"
-	qp.answers.PaddingLeft = 1
-	qp.answers.PaddingRight = 1
-	qp.answers.PaddingBottom = 1
-	qp.answers.Height = 14
-}
-
 func resetHandlers() {
 	termui.Handle("/sys/kbd", func(e termui.Event) {})
 	termui.Handle("/sys/kbd/s", func(e termui.Event) {})
 	termui.Handle("/sys/kbd/<enter>", func(e termui.Event) {})
 }
 
-func showFinalPage(q *interact.Question) {
-	instructions = termui.NewPar(
-		fmt.Sprintf(strings.Join([]string{"Thank you for taking the test",
-			"Your final score was %3.1f",
-			"We will get in touch with you soon."}, ". "), q.Totscore))
+func showFinalPage(msg string) {
+	instructions = termui.NewPar(msg)
 	instructions.BorderLabel = "Thank You"
 	instructions.Height = 10
 	instructions.Width = termui.TermWidth() / 2
@@ -218,41 +113,34 @@ func clear() {
 	termui.Body.Rows = termui.Body.Rows[:0]
 }
 
-func testType() string {
-	var testType string
-	if !demoTaken {
-		testType = "demo"
-	} else {
-		testType = "test"
-	}
-	return testType
+func finalScore() string {
+	return fmt.Sprintf(strings.Join([]string{"Thank you for taking the test",
+		"Your final score was %3.1f",
+		"We will get in touch with you soon."}, ". "))
 }
 
 func fetchAndDisplayQn() {
 	client := interact.NewGruQuizClient(conn)
 
 	q, err := client.GetQuestion(context.Background(),
-		&interact.Req{Repeat: false, Sid: sessionId, Token: *token,
-			TestType: testType()})
+		&interact.Req{Repeat: false, Sid: sessionId, Token: *token})
+
 	if err != nil {
 		log.Fatalf("Could not get question.Got err: %v", err)
 	}
 	curQuestion = q
 
 	// TODO(pawan) - If he has already taken the demo,don't show the screen again.
-	if q.Id == "END" {
+	if q.Id == "DEMOEND" {
 		clear()
-		if !demoTaken {
-			ts = 0.0
-			ls = 0.0
-			demoTaken = true
-			renderInstructionsPage()
-			return
-		}
-		showFinalPage(q)
+		ts = 0.0
+		ls = 0.0
+		renderInstructionsPage(true)
+		return
+	} else if q.Id == "END" {
+		showFinalPage(finalScore())
 		return
 	}
-
 	populateQuestionsPage(q)
 }
 
@@ -272,7 +160,8 @@ func streamRecv(stream interact.GruQuiz_StreamChanClient) {
 
 		if msg.Status == "END" {
 			clear()
-			showFinalPage(curQuestion)
+			showFinalPage("dummy")
+			// showFinalPage(curQuestion)
 			return
 		}
 		servTime, err = time.ParseDuration(msg.TimeLeft)
@@ -333,78 +222,8 @@ func initializeTest() {
 	if err := stream.Send(cliStat); err != nil {
 		log.Panic(err)
 	}
-
 	go streamRecv(stream)
 	go streamSend(stream)
-}
-
-func renderInstructionsPage() {
-	resetHandlers()
-	termui.Render(instructions)
-	// Adding an offset so that all these boxes come inside the instructions box.
-	termui.Body.Y = 2
-	termui.Body.AddRows(
-		termui.NewRow(
-			termui.NewCol(10, 1, infoPage.terminal)),
-		termui.NewRow(
-			termui.NewCol(10, 1, infoPage.general)),
-		termui.NewRow(
-			termui.NewCol(10, 1, infoPage.scoring)),
-		termui.NewRow(
-			termui.NewCol(10, 1, infoPage.contact)),
-		termui.NewRow(
-			termui.NewCol(10, 1, infoPage.demo)))
-
-	if demoTaken {
-		infoPage.demo.Text = "Press s to start the test."
-	}
-	termui.Body.Align()
-	termui.Render(termui.Body)
-
-	termui.Handle("/sys/kbd/s", func(e termui.Event) {
-		if !demoTaken {
-			initializeDemo()
-			return
-		}
-		clear()
-		initializeTest()
-	})
-}
-
-func renderQuestionsPage() {
-	termui.Body.Y = 0
-	termui.Body.AddRows(
-		termui.NewRow(
-			termui.NewCol(3, 0, qp.timeLeft),
-			termui.NewCol(3, 0, qp.timeSpent),
-			termui.NewCol(3, 0, qp.score),
-			termui.NewCol(3, 0, qp.lastScore)),
-		termui.NewRow(
-			termui.NewCol(10, 0, qp.que),
-			termui.NewCol(2, 0, qp.scoringInfo)),
-		termui.NewRow(
-			termui.NewCol(12, 0, qp.answers)))
-
-	termui.Body.Align()
-	termui.Render(termui.Body)
-
-	secondsCount := 0
-	lck.Lock()
-	leftTime = testDur * time.Minute
-	lck.Unlock()
-
-	termui.Handle("/timer/1s", func(e termui.Event) {
-		secondsCount += 1
-		timeTaken += 1
-		lck.Lock()
-		leftTime = leftTime - time.Second
-		lck.Unlock()
-		qp.timeSpent.Text = fmt.Sprintf("%02d:%02d", timeTaken/60,
-			timeTaken%60)
-		qp.timeLeft.Text = fmt.Sprintf("%02d:%02d", leftTime/time.Minute,
-			(leftTime%time.Minute)/time.Second)
-		termui.Render(termui.Body)
-	})
 }
 
 func renderSelectedAnswers(selected []string, m map[string]*interact.Answer) {
@@ -469,7 +288,7 @@ func enterHandler(e termui.Event, q *interact.Question, selected []string,
 			answerIds = append(answerIds, m[s].Id)
 		}
 		resp := interact.Response{Qid: q.Id, Aid: answerIds,
-			Sid: sessionId, Token: *token, TestType: testType()}
+			Sid: sessionId, Token: *token}
 		client := interact.NewGruQuizClient(conn)
 		client.SendAnswer(context.Background(), &resp)
 		fetchAndDisplayQn()
@@ -557,22 +376,32 @@ func populateQuestionsPage(q *interact.Question) {
 }
 
 func initializeDemo() {
-	// Set up a connection to the server.
-	var err error
-
-	client := interact.NewGruQuizClient(conn)
-	session, err := client.Authenticate(context.Background(), &interact.Token{Id: *token})
-	if err != nil {
-		infoPage.demo.Text = grpc.ErrorDesc(err) + " Press Ctrl+Q to exit and try again."
-		infoPage.demo.TextFgColor = termui.ColorRed
-		termui.Render(termui.Body)
-		return
-	}
-	sessionId = session.Id
 	clear()
 	setupQuestionsPage()
 	renderQuestionsPage()
 	fetchAndDisplayQn()
+}
+
+func setupInitialPage(s *interact.Session) {
+	state := s.State
+	if state == interact.Quiz_TEST_FINISHED {
+		//show final page saying test already taken and return
+		showFinalPage("You have already taken the test.")
+	}
+	sessionId = s.Id
+	if state == interact.Quiz_TEST_STARTED {
+		//call get questions and return
+	}
+	setupInfoPage(termui.TermHeight(), termui.TermWidth())
+	if state == interact.Quiz_DEMO_NOT_TAKEN {
+		//call instructions screen with demo taken to be false
+		renderInstructionsPage(false)
+	}
+	if state == interact.Quiz_TEST_NOT_TAKEN {
+		//call instructions screen with demo taken to be true
+		renderInstructionsPage(true)
+	}
+
 }
 
 func main() {
@@ -583,16 +412,34 @@ func main() {
 	}
 	defer termui.Close()
 
-	th := termui.TermHeight()
-	tw := termui.TermWidth()
-
 	conn, err = grpc.Dial(*address, grpc.WithInsecure())
 	if err != nil {
 		log.Fatal(err)
 		return
 	}
-	setupInfoPage(th, tw)
-	renderInstructionsPage()
+
+	instructions = termui.NewPar(
+		fmt.Sprintf("Wait a second, we are getting you started..."))
+	instructions.BorderLabel = "Authenticating"
+	instructions.Height = 10
+	instructions.Width = termui.TermWidth() / 2
+	instructions.Y = termui.TermHeight() / 4
+	instructions.X = termui.TermWidth() / 4
+	instructions.PaddingTop = 1
+	instructions.PaddingLeft = 1
+	termui.Render(instructions)
+
+	// Set up a connection to the server.
+
+	client := interact.NewGruQuizClient(conn)
+	s, err := client.Authenticate(context.Background(), &interact.Token{Id: *token})
+	if err != nil {
+		instructions.Text = grpc.ErrorDesc(err) + " Press Ctrl+Q to exit and try again."
+		instructions.TextFgColor = termui.ColorRed
+		termui.Render(instructions)
+		return
+	}
+	setupInitialPage(s)
 
 	// Pressing Ctrl-q terminates the ui.
 	termui.Handle("/sys/kbd/C-q", func(termui.Event) {
