+++ server/demo_test.yaml
@@ -45,4 +45,4 @@
   positive: 5
   negative: 2.5
   tags: [Easy,demo]
-  
+
