+++ server/logs/abcd1234.log
@@ -1,0 +1,2 @@
+2016/06/20 06:23:45 UTC session_token XVlBzgbaiCMRAjWwhTHctcuAxhxKQFDaFpLS
+2016/06/20 06:23:48 UTC ping demo-1
