+++ server/logs/test_token_copy.log
@@ -1,0 +1,7 @@
+2016/06/14 07:12:32 UTC session_token XVlBzgbaiCMRAjWwhTHctcuAxhxKQFDaFpLS
+2016/06/14 07:12:37 UTC test_start
+2016/06/14 07:12:38 UTC response demo-1 5.0
+2016/06/14 07:12:40 UTC ping demo-2
+2016/06/14 07:12:45 UTC ping demo-2
+2016/06/14 07:12:47 UTC response demo-2 demo-2a,demo-2b 10.0
+2016/06/14 07:12:50 UTC ping demo-3
